# src/lib/swagger/calculation-services.swagger.yaml
openapi: 3.0.0
info:
  title: EDK Calculation Services API
  description: API for managing energy calculation services for NYC building retrofits
  version: 1.0.0
  contact:
    name: EDK API Support
    email: support@edk.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://production-domain.com/api
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error description
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
          description: Validation error details
    ValidationError:
      type: object
      properties:
        field:
          type: string
          description: Field that failed validation
        message:
          type: string
          description: Validation error message
    ValidationWarning:
      type: object
      properties:
        field:
          type: string
          description: Field with warning
        message:
          type: string
          description: Warning message
...
    LL97CalculationData:
      type: object
      properties:
        calculationId:
          type: string
          format: uuid
        lastCalculated:
          type: string
          format: date-time
        serviceVersion:
          type: string
        emissionsBudget2024to2029:
          type: number
        emissionsBudget2030to2034:
          type: number
        emissionsBudget2035to2039:
          type: number
        emissionsBudget2040to2049:
          type: number
        totalBuildingEmissionsLL84:
          type: number
        annualFeeExceedingBudget2024to2029:
          type: number
        annualFeeExceedingBudget2030to2034:
          type: number
        annualFeeExceedingBudget2035to2039:
          type: number
        annualFeeExceedingBudget2040to2049:
          type: number
        beCreditBefore2027:
          type: number
        beCredit2027to2029:
          type: number
        adjustedTotalBuildingEmissions2024to2029:
          type: number
        adjustedTotalBuildingEmissions2030to2034:
          type: number
        adjustedTotalBuildingEmissions2035to2039:
          type: number
        adjustedTotalBuildingEmissions2040to2049:
          type: number
        adjustedAnnualFeeBefore2027:
          type: number
        adjustedAnnualFee2027to2029:
          type: number
        adjustedAnnualFee2030to2034:
          type: number
        adjustedAnnualFee2035to2039:
          type: number
        adjustedAnnualFee2040to2049:
          type: number
    LL97CalculationOverrides:
      type: object
      properties:
        buildingClass:
          type: string
        totalSquareFeet:
          type: number
          minimum: 1
        totalBuildingEmissionsLL84:
          type: number
          minimum: 0
        annualBuildingMMBtuHeatingPTAC:
          type: number
          minimum: 0
        annualBuildingkWhHeatingPTHP:
          type: number
          minimum: 0
        feePerTonCO2e:
          type: number
          minimum: 0
        efGas:
          type: number
          minimum: 0
        efGrid2024to2029:
          type: number
          minimum: 0
        efGrid2030to2034:
          type: number
          minimum: 0
        beCoefficientBefore2027:
          type: number
          minimum: 0
        beCoefficient2027to2029:
          type: number
          minimum: 0
...
    NOICalculationData:
      type: object
      properties:
        calculationId:
          type: string
          format: uuid
        lastCalculated:
          type: string
          format: date-time
        serviceVersion:
          type: string
        potentialRentalIncomeImpact:
          type: number
        utilitiesIncludedInRent:
          type: boolean
        rentIncreasePercentage:
          type: number
        operatingExpenseSavings:
          type: number
        noiImpact:
          type: number
        effectiveGrossIncomeChange:
          type: number
        netOperatingIncomeChange:
          type: number
        estimatedCurrentNOI:
          type: number
        newNOI:
          type: number
        noiYieldOnInvestment:
          type: number
        roiFromNOI:
          type: number
        noiPaybackPeriod:
          type: number
    NOICalculationOverrides:
      type: object
      properties:
        buildingValue:
          type: number
          minimum: 1
        capRate:
          type: number
          minimum: 0
        totalRetrofitCost:
          type: number
          minimum: 0
        annualEnergySavings:
          type: number
        rentIncreasePercentage:
          type: number
          minimum: 0
          maximum: 100
        utilitiesIncludedInRent:
          type: boolean
        operatingExpenseRatio:
          type: number
          minimum: 0
          maximum: 1
        vacancyRate:
          type: number
          minimum: 0
          maximum: 50
    PropertyValueCalculationData:
      type: object
      properties:
        calculationId:
          type: string
          format: uuid
        lastCalculated:
          type: string
          format: date-time
        serviceVersion:
          type: string
        noiBasedValueIncrease:
          type: number
        greenPremiumValueIncrease:
          type: number
        energyEfficiencyValueIncrease:
          type: number
        totalPropertyValueIncrease:
          type: number
        currentPropertyValue:
          type: number
        newPropertyValue:
          type: number
        valueCreationROI:
          type: number
        propertyAppreciationPercentage:
          type: number
        additionalLoanCapacity:
          type: number
        netBenefit:
          type: number
        valuePaybackPeriod:
          type: number
        refinancingBenefit:
          type: number
        dscrImprovement:
          type: number
        creditworthinessImprovement:
          type: number
    PropertyValueCalculationOverrides:
      type: object
      properties:
        currentPropertyValue:
          type: number
          minimum: 1
        totalRetrofitCost:
          type: number
          minimum: 0
        netOperatingIncomeChange:
          type: number
        energyReductionPercentage:
          type: number
          minimum: 0
          maximum: 100
        capRate:
          type: number
          minimum: 0
        greenPremiumPercentage:
          type: number
          minimum: 0
          maximum: 20
        energyEfficiencyPremium:
          type: number
          minimum: 0
          maximum: 20
        marketAppreciationRate:
          type: number
          minimum: 0
          maximum: 20
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
        calculationId:
          type: string
          format: uuid
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/ValidationWarning'
paths:
...
  /calculations/{calculationId}/ll97:
    get:
      summary: Get LL97 calculation data
      description: Retrieve Local Law 97 compliance calculation results
      tags: [LL97 Calculations]
      parameters:
        - name: calculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: LL97 calculation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LL97CalculationData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update LL97 calculation with overrides
      description: Update LL97 calculation parameters and cascade to dependent services
      tags: [LL97 Calculations]
      parameters:
        - name: calculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LL97CalculationOverrides'
      responses:
        '200':
          description: Calculation updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /calculations/{calculationId}/noi:
    get:
      summary: Get NOI calculation data
      description: Retrieve Net Operating Income impact analysis results
      tags: [NOI Calculations]
      parameters:
        - name: calculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: NOI calculation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NOICalculationData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update NOI calculation with overrides
      description: Update NOI calculation parameters and cascade to dependent services
      tags: [NOI Calculations]
      parameters:
        - name: calculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NOICalculationOverrides'
      responses:
        '200':
          description: Calculation updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /calculations/{calculationId}/property-value:
    get:
      summary: Get property value calculation data
      description: Retrieve property value impact analysis results
      tags: [Property Value Calculations]
      parameters:
        - name: calculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Property value calculation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyValueCalculationData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update property value calculation with overrides
      description: Update property value calculation parameters (final service, no cascading)
      tags: [Property Value Calculations]
      parameters:
        - name: calculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyValueCalculationOverrides'
      responses:
        '200':
          description: Calculation updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  responses:
    Unauthorized:
      description: Unauthorized - valid JWT token required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...